% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCADIE_core.R
\name{Estimate_sd_general}
\alias{Estimate_sd_general}
\title{Standard Error Estimation for SCADIE through Jackknife (or Bootstrap)}
\usage{
Estimate_sd_general(
  input_list,
  update_W_method = "SCAD",
  method = "jackknife",
  cores,
  bs_num = 500,
  H_update_method = "NNLS",
  H_update_gene = "all",
  signature_gene_row_index = NULL,
  jk_subsample = NA,
  max_itr = 50,
  weight_update_number = 999,
  cutoff = 10^-6,
  eta_weight = 4
)
}
\arguments{
\item{input_list}{A list containing the following fields: bulk_full_1,bulk_full_2, initial_H_1, initial_H_2, initial_W_1, initial_W_2, bulk_sub_1, bulk_sub_2. For a detailed explanation of these variables, please refer to the document of ``Iterate_W_H_full_general"}

\item{update_W_method}{The method use for W-update, can be "SCAD" or "NNLS", default ``SCAD"}

\item{method}{\verb{jackknife" or }bootstrap", default ``jackknife"}

\item{cores}{Number of cores used for computating}

\item{bs_num}{Number of bootstrap re-sampling, it is recommended to use at least 2x sample size}

\item{H_update_method}{the $W$-update method, can be \verb{NNLS" or }DWLS", default is ``NNLS"}

\item{H_update_gene}{in the $H$-update step, what set of genes to used for regression, this can be \verb{all" for using full $Y$ and $W$, or }signature" for using only those signature genes, default is ``signature"}

\item{signature_gene_row_index}{a vector containing signature genes' row indexes in full $Y$ or $W$}

\item{jk_subsample}{Number of jackknife resamplings, the default strategy for jackknife is to exhaustively apply levae-one-out through all samples, the only case we need to specify this argument is when sample size is extrmely large. Default value is NA, which equals the smaller number of samples in group1 and 2.}

\item{max_itr}{max iteration number, default 50.}

\item{weight_update_number}{update the weight matrix every \verb{weight_update_number" rounds. Noted that this is a new feature that has not been included in our paper (default 999, which does not update $E$ during iteration). In the current set-up, the weight matrix $E$ is obtained from initial }warm-up" runs. But to make it more data-adaptive, we can update $E$ during the iteration.}

\item{cutoff}{the cutoff coefficient for algorithm convergence, the algorithm will stop when the norm difference between previous $W$ and current $W$ is smaller than cutoff*W_norm, default $10^-8$}

\item{eta_weight}{$\eta_{n}$ for SCAD penalty}
}
\value{
When method=\verb{jackknife", the output is a list list of 4 fields: W1_vec, W2_vec,  w_diff_sd_jackknife and w_diff_sd_jackknife_raw. The W1_vec and W2_vec record all output W1s, W2s throughout the leave-one-out process. w_diff_sd_jackknife_raw.records the sample sd from all resamplings, and w_diff_sd_jackknife is the entry-wise sd. For DEG analysis, w_diff_sd_jackknife should be used. When method=}bootstrap", there would not be w_diff_sd_jackknife_raw.records in the output.
}
\description{
This function estiamtes the entry-wise standard error for W2-W1, which is used in combined with point estimates from `Iterate_W_H_full_general" for DEG identification. The default strategy is through leave-one-out Jackknife, but bootstrap is also implemented.
}
\examples{
Please refer to package vignette for a detailed walkthrough.
}
