% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCADIE_core.R
\name{Iterate_W_H_full_general}
\alias{Iterate_W_H_full_general}
\title{Main function for SCADIE point estimate}
\usage{
Iterate_W_H_full_general(
  c = NA,
  n_ct,
  input_initial_H1,
  input_initial_H2,
  input_initial_W1,
  input_initial_W2,
  input_bulk_sub1,
  input_bulk_sub2,
  input_bulk_full1,
  input_bulk_full2,
  update_W_method = "SCAD",
  var_estimate = NA,
  H_update_method = "NNLS",
  H_update_gene = "all",
  signature_gene_row_index,
  max_itr = 50,
  weight_update_number = 999,
  cutoff = 10^-6,
  eta_weight = 4
)
}
\arguments{
\item{c}{an auxicilliary argument used only when this function is called by ``Estimate_sd_general", default value is NA and should be kept like that.}

\item{n_ct}{number of cell types}

\item{input_initial_H1}{initial cell type proportion matrix $H_{1}$, each row represents a cell type and each column represents a sample}

\item{input_initial_H2}{initial cell type proportion matrix $H_{2}$, each row represents a cell type and each column represents a sample}

\item{input_initial_W1}{initial full $W_1$ matrix. Each row represents a gene and each column represents a cell type. Can be obtained by the ``update_W" function with $Y$ and initial $H$ as input. Please be noted that the $W$ is required to have rownames that are each gene's name}

\item{input_initial_W2}{initial full $W_2$ matrix. Each row represents a gene and each column represents a cell type. Can be obtained by the ``update_W" function with $Y$ and initial $H$ as input. Please be noted that the $W$ is required to have rownames that are each gene's name}

\item{input_bulk_sub1}{initial sub $Y_1$ matrix containing only the signature genes. Each row represents a gene and each column represents a sample.}

\item{input_bulk_sub2}{initial sub $Y_2$ matrix containing only the signature genes. Each row represents a gene and each column represents a sample.}

\item{input_bulk_full1}{initial full bulk $Y_1$ matrix. Each row represents a gene and each column represents a sample.}

\item{input_bulk_full2}{initial full bulk $Y_2$ matrix. Each row represents a gene and each column represents a sample.}

\item{update_W_method}{the $W$-update method, can be \verb{NNLS" or }SCAD", default is ``SCAD"}

\item{var_estimate}{an auxcilliary argument used when called by `Estimate_sd_general", default is NA and no need to change it when only running this function}

\item{H_update_method}{the $W$-update method, can be \verb{NNLS" or }DWLS", default is ``NNLS"}

\item{H_update_gene}{in the $H$-update step, what set of genes to used for regression, this can be \verb{all" for using full $Y$ and $W$, or }signature" for using only those signature genes, default is ``signature"}

\item{signature_gene_row_index}{a vector containing signature genes' row indexes in full $Y$ or $W$}

\item{max_itr}{max iteration number, default 50.}

\item{weight_update_number}{update the weight matrix every \verb{weight_update_number" rounds. Noted that this is a new feature that has not been included in our paper. In the current set-up, the weight matrix $E$ is obtained from initial }warm-up" runs. But to make it more data-adaptive, we can update $E$ during the iteration.}

\item{cutoff}{the cutoff coefficient for algorithm convergence, the algorithm will stop when the norm difference between previous $W$ and current $W$ is smaller than cutoff*W_norm, default $10^-8$}

\item{eta_weight}{$\eta_{n}$ for SCAD penalty}
}
\value{
a list with length two, the first list field contains all the outputs related to group 1, including end $W_{1}, H_{1}$, the weight matrix $E$, the updated matrix $E$ (if any), as well as a record of initial $W_{1}$; the second list field contains all the outputs related to group 2, including end $W_{2}, H_{2}$, the weight matrix $E$, the updated matrix $E$ (if any), as well as a record of initial $W_{2}$
}
\description{
This is the main interface function for users. Due to the large number of arguments it take in (W1, W2, H1, H2, etc.), it is recommended to wrap them all in a list. As the list will be extremely useful also in the standard error estimate function ``Estimate_sd_general"
}
\examples{
Please refer to package vignette for a detailed walkthrough.
}
