% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCADIE_core.R
\name{SIR_itr_general}
\alias{SIR_itr_general}
\title{Implementation of separate W-update without SCAD Penalty}
\usage{
SIR_itr_general(
  ini_H_adjusted,
  ini_W_adjusted,
  bulk_expr_sub,
  bulk_expr_full,
  n_ct,
  itr = 200,
  H_update_method = "NNLS",
  H_update_gene = "all",
  signature_gene_row_index,
  duplicated_rows = F
)
}
\arguments{
\item{ini_H_adjusted}{Initial H}

\item{ini_W_adjusted}{Initial W}

\item{bulk_expr_sub}{Bulk Y with only siganture gene rows}

\item{bulk_expr_full}{Full bulk Y}

\item{n_ct}{Number of cell types}

\item{itr}{Number of iteration}

\item{H_update_method}{\verb{NNLS" or }DWLS"}

\item{H_update_gene}{\verb{all" or }signature"}

\item{signature_gene_row_index}{row indexes of signature genes in full W or Y}

\item{duplicated_rows}{If there are duplicated rows present in Y}
}
\value{
Returns a list of W and H after iteration, as well as initial W and H
}
\description{
This function should be called by parental function and is not supposed to be used individually.
}
