% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCADIE_core.R
\name{SIR_itr_W_sim_updated_general}
\alias{SIR_itr_W_sim_updated_general}
\title{Implementation of SCAD-penalty itertion
This is a sub-function and is supposed to be called by its parental function}
\usage{
SIR_itr_W_sim_updated_general(
  ini_H_adjusted1,
  ini_H_adjusted2,
  ini_W_adjusted1,
  ini_W_adjusted2,
  bulk_expr_sub1,
  bulk_expr_sub2,
  bulk_expr_full1,
  bulk_expr_full2,
  itr = 200,
  lambda,
  Weight_mat,
  H_update_method,
  H_update_gene,
  signature_gene_row_index,
  weight_update_number,
  cutoff,
  lambda_weight
)
}
\arguments{
\item{ini_H_adjusted1}{Initial H1}

\item{ini_H_adjusted2}{Initial H2}

\item{ini_W_adjusted1}{Initial W1}

\item{ini_W_adjusted2}{Initial W2}

\item{bulk_expr_sub1}{Bulk Y1 with only signature gene rows}

\item{bulk_expr_sub2}{Bulk Y2 with only signature gene rows}

\item{bulk_expr_full1}{Bulk Y1}

\item{bulk_expr_full2}{Bulk Y2}

\item{itr}{number of iteration}

\item{lambda}{the $\lambda$ controlling the penalty term in objective function}

\item{Weight_mat}{input weight matrix}

\item{H_update_method}{\verb{NNLS" or }DWLS"}

\item{H_update_gene}{\verb{siganture" or }all"}

\item{signature_gene_row_index}{signature genes' row indexes in full Y or W}

\item{weight_update_number}{update E every n times, not used}

\item{cutoff}{default 10^-6, see the document of ``Iterate_W_H_full_general" for more explanation}

\item{lambda_weight}{see the document of ``Iterate_W_H_full_general" for more explanation}
}
\value{
End Ws and Hs, initial Ws and Hs, as well as weight matrix E
}
\description{
Implementation of SCAD-penalty itertion
This is a sub-function and is supposed to be called by its parental function
}
