% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCADIE_core.R
\name{update_W_sim_adaptive1_general}
\alias{update_W_sim_adaptive1_general}
\title{Implementation of SCAD-penalized W update}
\usage{
update_W_sim_adaptive1_general(H1, H2, lambda, Y1, Y2, Weight_mat)
}
\arguments{
\item{H1}{cell type proportion matrix for group1}

\item{H2}{cell type proportion matrix for group2}

\item{lambda}{$\lambda$ controllling the SCAD penalty term in the objective function}

\item{Y1}{full bulk matrix Y1}

\item{Y2}{full bulk matrix Y2}

\item{Weight_mat}{weight matrix E}
}
\value{
the updated W1 and W2
}
\description{
Implementation of SCAD-penalized W update
}
